requirements.txt
Flask==0.11
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==1.0.3
Werkzeug==0.10.4
blinker==1.3
itsdangerous==0.24
wsgiref==0.1.2
psycopg2
py-bcrypt